requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.1.5
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
csv342==1.0.1
decorator==4.4.2
defusedxml==0.6.0
Eel==0.14.0
entrypoints==0.3
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.1
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
PyQt5==5.13.2
PyQt5-sip==12.8.1
pyqt5-tools==5.13.2.1.6rc1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytube3==9.6.4
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
youtube-dl==2020.9.14
zope.event==4.4
zope.interface==5.1.0
